refactor(cli): extract shared core to cli/_app.py (#284, step 2)
Move the Typer `app`, the root `@app.callback()`, `console`, and the helpers
every command uses (_get_store, _safe_exc, _inference_hint, _record_miss_event,
_build_miss_hint, _profile_from_ctx, _load_labeled_queries_jsonl,
_version_callback) into vstash/cli/_app.py, so the command-group modules
extracted next can `from ._app import app, console, _get_store, ...` without
importing the package __init__ (which would be a cycle). Bodies moved
byte-for-byte (verified via diff).

__init__.py imports what the commands use from ._app and drops the now-unused
imports (atexit, Console, VstashConfig, __version__) whose users moved.

Tests that mock _get_store's collaborators (load_config / atexit /
open_store_for_config) now patch vstash.cli._app, where _get_store reads those
bindings; _build_miss_hint's unit test imports it from cli._app. Behaviour is
unchanged.

- pytest tests/ -> 1290 passed; ruff clean; CLI commands resolve via CliRunner
